git remote: "origin" is not always the remote
Adding a top level flag for `pulls` and `issues` to specify an alternate
git remote (default is still "origin"). This took a little rearranging
of the Manager code, to be inside the app.Action (so the flags are
parsed at that point).

Now you can `pulls --remote crosbymichael` to see those instead of the
dotcloud list.

Docker-DCO-1.1-Signed-off-by: Vincent Batts <vbatts@redhat.com> (github: vbatts)